requirements.txt
pdfplumber==0.11.3
pandas==2.2.2
streamlit==1.37.0
openai==1.41.0
python-dotenv==1.0.0
aiohttp==3.10.5
aiofiles==24.1.0
nest-asyncio==1.6.0
